backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.23.0
httpx>=0.24.0
python-jose>=3.3.0
supabase>=2.3.0
pydantic>=1.10.0
